pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "requests",
    "python-dotenv",
    "mcp>=1.13.1",
]

[tool.uv]

